Moving the DTL Validation into the validate package
