+++ azurerm/data_source_dev_test_lab.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 
 	"github.com/hashicorp/terraform/helper/schema"
-	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/helpers/azure"
+	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/helpers/validate"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/utils"
 )
 
@@ -16,7 +16,7 @@ func dataSourceArmDevTestLab() *schema.Resource {
 			"name": {
 				Type:         schema.TypeString,
 				Required:     true,
-				ValidateFunc: azure.ValidateDevTestLabName(),
+				ValidateFunc: validate.DevTestLabName(),
 			},
 
 			"location": locationForDataSourceSchema(),
